Firestore integration

+++ config.py
@@ -1,9 +1,13 @@
 import json
 import firebase_admin
 from firebase_admin import credentials, firestore
-from os import path
+import os
 
-if path.exists('firebaseServiceAccountKey.json'):
+def is_running_on_gcp():
+    return os.getenv('FUNCTION_TARGET') is not None
+
+#Manage settings with Firestore only if running on GCP
+if is_running_on_gcp():
     cred = credentials.Certificate('firebaseServiceAccountKey.json')
     firebase_admin.initialize_app(cred)
     db = firestore.client()
@@ -23,7 +27,8 @@ class Config:
         self.google_functions_url = data.get("google_functions_url")
 
 def getConfig() -> Config:
-    if path.exists('firebaseServiceAccountKey.json'):
+    
+    if is_running_on_gcp():
         config_data = doc_ref.get().to_dict()
 
     else:
@@ -33,21 +38,28 @@ def getConfig() -> Config:
     config = Config(config_data)
     return config
 
+
 def get(parameter: str):
-    if path.exists('firebaseServiceAccountKey.json'):
+    
+    if is_running_on_gcp():
         data = doc_ref.get().to_dict()
+    
     else:
         with open('config.json') as file:
             data = json.load(file)
+    
     return data[parameter]
 
+
 def write(parameter: str, value):
-    if path.exists('firebaseServiceAccountKey.json'):
+    
+    if is_running_on_gcp():
         doc_ref.update({parameter: value})
 
-    with open('config.json', 'r+') as file:
-        data = json.load(file)
-        data[parameter] = value
-        file.seek(0)
-        json.dump(data, file, indent=4)
-        file.truncate()
+    else:
+        with open('config.json', 'r+') as file:
+            data = json.load(file)
+            data[parameter] = value
+            file.seek(0)
+            json.dump(data, file, indent=4)
+            file.truncate()
